refactor: move UI sections to layouts folder
